Move logger code to src/client/logging/.
